pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "blamebug"
version = "0.1.0"
description = "Automated incident log analysis (webhook + LLM + Gradio)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "gradio>=4.44.0",
  "huggingface_hub>=0.19.0,<0.26.0",
  "openai>=1.12.0",
  "httpx>=0.26.0",
  "pydantic>=2.5.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["blamebug*"]
